requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
discord==2.3.2
discord.py==2.5.2
exceptiongroup==1.3.0
executing==2.2.0
frozenlist==1.6.0
h11==0.14.0
httpcore==0.17.3
idna==3.10
multidict==6.4.3
propcache==0.3.1
Pygments==2.19.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
youtube-search-python==1.6.6
